PoP: Fixed the logical error in level skip category detection
So basically, when you managed to press Shift + L on the very first frame, the level skip category detection failed due to a logical error I had made earlier.
Instead of checking if FrameSeconds is not equal 718 while also the minutes is 15, I put them as separate conditions. Due to this when someone pressed Shift + L on the first frame possible, FrameSeconds is equal to 718 and hence it didn't activate the level skip condition despite the minute being 60 rather than 15. This is my bad, sorry for not being frame perfect during testing earlier.