Fix memory leak in TFS
We were subscribing to the game.Exited event but we never unsubscribed, which meant that when we closed the game the Process object could not be garbage collected, thus every game restart would cause a memory leak in a new Process object. Although we never noticed, it is likely that 100 or more game restarts would have caused a performance issue and possibly even a bug in the script.